fix missing validation when setting retention on topic level.
